+++ src/features/commands/infoImporter.ts
@@ -5,6 +5,7 @@ import * as xmldoc from 'xmldoc';
 import ProppedModel from '../../other/proppedModel';
 import AnnoXml from '../../other/annoXml';
 import * as channel from '../channel';
+import { Quaternion } from '../../other/math';
 
 export class InfoImporter {
   insertPropContainer: xmldoc.XmlElement | undefined = undefined;
@@ -25,17 +26,21 @@ export class InfoImporter {
 
         let openUri = await vscode.window.showOpenDialog(options);
         if (openUri && openUri[0]) {
-          channel.show();
-          channel.log(`Import from ${path.basename(openUri[0].fsPath)} into ${path.basename(fileUri.fsPath)}`);
-
-          const importer = new InfoImporter();
-          importer.importInfo(fileUri.fsPath, openUri[0].fsPath);
+          InfoImporter.commandImportInfo(fileUri.fsPath, openUri[0].fsPath);
         }
       })
     ];
 
     return disposable;
 	}
+
+  public static commandImportInfo(cfgFilePath: string, gltfFilePath: string) {
+    channel.show();
+    channel.log(`Import from ${path.basename(gltfFilePath)} into ${path.basename(cfgFilePath)}`);
+
+    const importer = new InfoImporter();
+    importer.importInfo(cfgFilePath, gltfFilePath);
+  }
 
   public importInfo(targetFile: string, modelFile: string) {
     const model = ProppedModel.fromFile(modelFile);
@@ -48,6 +53,33 @@ export class InfoImporter {
     }
     else {
       channel.log('No \'ground\' node/mesh found. Skip BuildBlocker');
+    }
+
+    const hitboxes = model.getHitBoxes();
+    if (hitboxes && hitboxes.length > 0) {
+      xml.ensureSection('Info', [ { } ]);
+      channel.log(`${hitboxes.length} hitboxes found`);
+
+      xml.remove('//Info/IntersectBox', true);
+      for (let hitbox of hitboxes) {
+        const parent = xml.findElement('//Info'); 
+        if (parent) {
+          /* eslint-disable @typescript-eslint/naming-convention */
+          parent.createChild('IntersectBox', { after: [ 'IntersectBox', 'DisableFeedbackArea', 'MeshBoundingBox', 'BoundingBox' ] }).set({
+            Name: hitbox.name,
+            Position: hitbox.center.toFixedF(),
+            Rotation: Quaternion.default.toFixedF(0),
+            Extents: hitbox.size.div(2).toFixedF()
+          });
+          /* eslint-enable @typescript-eslint/naming-convention */
+        }
+        else {
+          console.error(`ensureSection should have created //Info`);
+        }
+      }
+    }
+    else {
+      channel.log('No node/mesh starting with \'hitbox\' found. Skip hitboxes');
     }
 
     fs.writeFileSync(targetFile, xml.toString());