+++ src/test/suite/import.test.ts
@@ -5,6 +5,7 @@ import * as utils from '../../other/utils';
 import * as xml2js from 'xml2js';
 
 import { PropImporter } from '../../features/commands/propImporter';
+import { InfoImporter } from '../../features/commands/infoImporter';
 
 suite('import tests', () => {
   // clear to avoid old files leading to wrong results
@@ -22,6 +23,20 @@ suite('import tests', () => {
 
     const imported = await xml2js.parseStringPromise(fs.readFileSync(simpleCfg, 'utf8'));
     const expected = await xml2js.parseStringPromise(fs.readFileSync('../../src/test/suite/data/simple-expected.cfg', 'utf8'));
+
+    assert.deepStrictEqual(imported, expected);
+  });
+
+  test('hitbox import', async () => {
+    const propsGltf = path.resolve('../../src/test/suite/data/props.gltf');
+    const ifo = path.resolve('../../out/test/suite/data/hitbox.ifo');
+    utils.ensureDir(path.dirname(ifo));
+    fs.copyFileSync('../../src/test/suite/data/hitbox.ifo', ifo);
+
+    InfoImporter.commandImportInfo(ifo, propsGltf);
+
+    const imported = await xml2js.parseStringPromise(fs.readFileSync(ifo, 'utf8'));
+    const expected = await xml2js.parseStringPromise(fs.readFileSync('../../src/test/suite/data/hitbox-expected.ifo', 'utf8'));
 
     assert.deepStrictEqual(imported, expected);
   });